Move Rx2 to rx module
