+++ libraries/rib-rx/src/main/java/com/badoo/ribs/rx/adapter/Rx2.kt
@@ -1,5 +1,6 @@
-package com.badoo.ribs.core.state
+package com.badoo.ribs.rx.adapter
 
+import com.badoo.ribs.core.state.Source
 import io.reactivex.Observable
 
 fun <T> Source<T>.rx2(): Observable<T> =
