+++ sandbox/src/main/java/com/badoo/ribs/sandbox/rib/switcher/SwitcherInteractor.kt
@@ -9,7 +9,6 @@ import com.badoo.ribs.android.dialog.Dialog
 import com.badoo.ribs.clienthelper.interactor.Interactor
 import com.badoo.ribs.core.Node
 import com.badoo.ribs.core.modality.BuildParams
-import com.badoo.ribs.core.state.rx2
 import com.badoo.ribs.routing.router.Router
 import com.badoo.ribs.routing.router.Router.TransitionState.IN_TRANSITION
 import com.badoo.ribs.routing.router.Router.TransitionState.SETTLED
@@ -17,6 +16,7 @@ import com.badoo.ribs.routing.source.backstack.BackStack
 import com.badoo.ribs.routing.source.backstack.operation.pop
 import com.badoo.ribs.routing.source.backstack.operation.push
 import com.badoo.ribs.routing.source.backstack.operation.pushOverlay
+import com.badoo.ribs.rx.adapter.rx2
 import com.badoo.ribs.sandbox.rib.blocker.Blocker
 import com.badoo.ribs.sandbox.rib.menu.Menu
 import com.badoo.ribs.sandbox.rib.menu.Menu.Input.SelectMenuItem
